brand-spy: fix tier_score NUMERIC(6,3) overflow for large brands
Root cause: tier_score = poolSize - rank + 1. For norseorganics.co
(1289 active ads), top-ranked ad gets score 1289, which overflows
NUMERIC(6,3) (max 999.999). This caused scoreBrand to fail silently
for ANY brand with 1000+ active ads.

Fix: migration 045 widens the column to INTEGER. Update cast in
scoreBrand bulk UPDATE to match (::numeric → ::integer).

+++ server/src/workers/brandSpyWorker.js
@@ -170,7 +170,7 @@ export async function scoreBrand(brandId) {
     const AD_UPDATE_COLS = 10; // id, cr, r3, r7, r21, v7d, v21d, ps, tier, ts
     const updatePlaceholders = ranked.map((_, i) => {
       const b = i * AD_UPDATE_COLS + 1;
-      return `($${b}::uuid,$${b+1}::integer,$${b+2}::integer,$${b+3}::integer,$${b+4}::integer,$${b+5}::integer,$${b+6}::integer,$${b+7}::integer,$${b+8}::text,$${b+9}::numeric)`;
+      return `($${b}::uuid,$${b+1}::integer,$${b+2}::integer,$${b+3}::integer,$${b+4}::integer,$${b+5}::integer,$${b+6}::integer,$${b+7}::integer,$${b+8}::text,$${b+9}::integer)`;
     });
     const updateParams = ranked.flatMap((r) => {
       const hist = historical.get(r.adArchiveId) ?? { d3: null, d7: null, d21: null };
